upt: improved task splitter
