+++ tests/test_chat.py
@@ -5,7 +5,7 @@ import logging
 from langsmith import traceable
 from pydantic import BaseModel, Field
 
-from grafo.interpreters import LLM, OpenAIHandler
+from grafo.handlers import LLM, OpenAIHandler
 from grafo._internal import logger
 
 # disable loggers
